+++ packages/demo-saas-next12/middleware.ts
@@ -1,16 +1,29 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { getSession } from '@frontegg/nextjs/edge';
+import { getSession, shouldByPassMiddleware } from '@frontegg/nextjs/edge';
 
 export const middleware = async (request: NextRequest) => {
-  const session = await getSession(request);
-  const isAuthRoute = request.url.endsWith('/') || request.url.endsWith('/force-session');
-
-  if (!session && isAuthRoute) {
-    // redirect unauthenticated user to /account/login page
-    return NextResponse.redirect(new URL('/account/login', request.url));
+  // this if for frontegg middleware tests
+  if (process.env['FRONTEGG_TEST_URL']) {
+    return NextResponse.next();
   }
 
+  const pathname = request.nextUrl.pathname;
+
+  if (
+    shouldByPassMiddleware(pathname, {
+      bypassImageOptimization: false,
+    })
+  ) {
+    return NextResponse.next();
+  }
+
+  const session = await getSession(request);
+  if (!session) {
+    //  redirect unauthenticated user to /account/login page
+    const loginUrl = `/account/login?redirectUrl=${encodeURIComponent(pathname)}`;
+    return NextResponse.redirect(new URL(loginUrl, process.env['FRONTEGG_APP_URL']));
+  }
   return NextResponse.next();
 };