+++ packages/demo-saas-next12/pages/_app.tsx
@@ -9,4 +9,7 @@ function CustomApp({ Component, pageProps }: AppProps) {
 export default withFronteggApp(CustomApp, {
   hostedLoginBox: true,
   customLoader: true,
+  authOptions: {
+    keepSessionAlive: true,
+  },
 });